Move visitors back to GraphQLService because of variables
