+++ src/RequestLoader.cpp
@@ -8,277 +8,14 @@
 #include <array>
 #include <sstream>
 
-namespace graphql::runtime {
+namespace graphql::generator {
 
-using namespace service;
-
-ValueVisitor::ValueVisitor(const response::Value& variables)
-	: _variables(variables)
+RequestLoader::RequestLoader()
 {
 }
 
-response::Value ValueVisitor::getValue()
-{
-	auto result = std::move(_value);
-
-	return result;
-}
-
-void ValueVisitor::visit(const peg::ast_node& value)
-{
-	if (value.is_type<peg::variable_value>())
-	{
-		visitVariable(value);
-	}
-	else if (value.is_type<peg::integer_value>())
-	{
-		visitIntValue(value);
-	}
-	else if (value.is_type<peg::float_value>())
-	{
-		visitFloatValue(value);
-	}
-	else if (value.is_type<peg::string_value>())
-	{
-		visitStringValue(value);
-	}
-	else if (value.is_type<peg::true_keyword>() || value.is_type<peg::false_keyword>())
-	{
-		visitBooleanValue(value);
-	}
-	else if (value.is_type<peg::null_keyword>())
-	{
-		visitNullValue(value);
-	}
-	else if (value.is_type<peg::enum_value>())
-	{
-		visitEnumValue(value);
-	}
-	else if (value.is_type<peg::list_value>())
-	{
-		visitListValue(value);
-	}
-	else if (value.is_type<peg::object_value>())
-	{
-		visitObjectValue(value);
-	}
-}
-
-void ValueVisitor::visitVariable(const peg::ast_node& variable)
-{
-	const std::string name(variable.string_view().substr(1));
-	auto itr = _variables.find(name);
-
-	if (itr == _variables.get<response::MapType>().cend())
-	{
-		auto position = variable.begin();
-		std::ostringstream error;
-
-		error << "Unknown variable name: " << name;
-
-		throw schema_exception {
-			{ schema_error { error.str(), { position.line, position.column } } }
-		};
-	}
-
-	_value = response::Value(itr->second);
-}
-
-void ValueVisitor::visitIntValue(const peg::ast_node& intValue)
-{
-	_value = response::Value(std::atoi(intValue.string().c_str()));
-}
-
-void ValueVisitor::visitFloatValue(const peg::ast_node& floatValue)
-{
-	_value = response::Value(std::atof(floatValue.string().c_str()));
-}
-
-void ValueVisitor::visitStringValue(const peg::ast_node& stringValue)
-{
-	_value = response::Value(std::string { stringValue.unescaped_view() });
-}
-
-void ValueVisitor::visitBooleanValue(const peg::ast_node& booleanValue)
-{
-	_value = response::Value(booleanValue.is_type<peg::true_keyword>());
-}
-
-void ValueVisitor::visitNullValue(const peg::ast_node& /*nullValue*/)
-{
-	_value = {};
-}
-
-void ValueVisitor::visitEnumValue(const peg::ast_node& enumValue)
-{
-	_value = response::Value(response::Type::EnumValue);
-	_value.set<response::StringType>(enumValue.string());
-}
-
-void ValueVisitor::visitListValue(const peg::ast_node& listValue)
-{
-	_value = response::Value(response::Type::List);
-	_value.reserve(listValue.children.size());
-
-	ValueVisitor visitor(_variables);
-
-	for (const auto& child : listValue.children)
-	{
-		visitor.visit(*child);
-		_value.emplace_back(visitor.getValue());
-	}
-}
-
-void ValueVisitor::visitObjectValue(const peg::ast_node& objectValue)
-{
-	_value = response::Value(response::Type::Map);
-	_value.reserve(objectValue.children.size());
-
-	ValueVisitor visitor(_variables);
-
-	for (const auto& field : objectValue.children)
-	{
-		visitor.visit(*field->children.back());
-		_value.emplace_back(field->children.front()->string(), visitor.getValue());
-	}
-}
-
-DirectiveVisitor::DirectiveVisitor(const response::Value& variables)
-	: _variables(variables)
-	, _directives(response::Type::Map)
+void RequestLoader::visit(const peg::ast& request, std::string_view operationName)
 {
 }
 
-void DirectiveVisitor::visit(const peg::ast_node& directives)
-{
-	response::Value result(response::Type::Map);
-
-	for (const auto& directive : directives.children)
-	{
-		std::string directiveName;
-
-		peg::on_first_child<peg::directive_name>(*directive,
-			[&directiveName](const peg::ast_node& child) {
-				directiveName = child.string_view();
-			});
-
-		if (directiveName.empty())
-		{
-			continue;
-		}
-
-		response::Value directiveArguments(response::Type::Map);
-
-		peg::on_first_child<peg::arguments>(*directive,
-			[this, &directiveArguments](const peg::ast_node& child) {
-				ValueVisitor visitor(_variables);
-
-				for (auto& argument : child.children)
-				{
-					visitor.visit(*argument->children.back());
-
-					directiveArguments.emplace_back(argument->children.front()->string(),
-						visitor.getValue());
-				}
-			});
-
-		result.emplace_back(std::move(directiveName), std::move(directiveArguments));
-	}
-
-	_directives = std::move(result);
-}
-
-response::Value DirectiveVisitor::getDirectives()
-{
-	auto result = std::move(_directives);
-
-	return result;
-}
-
-bool DirectiveVisitor::shouldSkip() const
-{
-	static const std::array<std::pair<bool, std::string>, 2> skippedNames = {
-		std::make_pair<bool, std::string>(true, "skip"),
-		std::make_pair<bool, std::string>(false, "include"),
-	};
-
-	for (const auto& entry : skippedNames)
-	{
-		const bool skip = entry.first;
-		auto itrDirective = _directives.find(entry.second);
-
-		if (itrDirective == _directives.end())
-		{
-			continue;
-		}
-
-		auto& arguments = itrDirective->second;
-
-		if (arguments.type() != response::Type::Map)
-		{
-			std::ostringstream error;
-
-			error << "Invalid arguments to directive: " << entry.second;
-
-			throw schema_exception { { error.str() } };
-		}
-
-		bool argumentTrue = false;
-		bool argumentFalse = false;
-
-		for (auto& argument : arguments)
-		{
-			if (argumentTrue || argumentFalse || argument.second.type() != response::Type::Boolean
-				|| argument.first != "if")
-			{
-				std::ostringstream error;
-
-				error << "Invalid argument to directive: " << entry.second
-					  << " name: " << argument.first;
-
-				throw schema_exception { { error.str() } };
-			}
-
-			argumentTrue = argument.second.get<response::BooleanType>();
-			argumentFalse = !argumentTrue;
-		}
-
-		if (argumentTrue)
-		{
-			return skip;
-		}
-		else if (argumentFalse)
-		{
-			return !skip;
-		}
-		else
-		{
-			std::ostringstream error;
-
-			error << "Missing argument directive: " << entry.second << " name: if";
-
-			throw schema_exception { { error.str() } };
-		}
-	}
-
-	return false;
-}
-
-FragmentDefinitionVisitor::FragmentDefinitionVisitor(const response::Value& variables)
-	: _variables(variables)
-{
-}
-
-FragmentMap FragmentDefinitionVisitor::getFragments()
-{
-	FragmentMap result(std::move(_fragments));
-	return result;
-}
-
-void FragmentDefinitionVisitor::visit(const peg::ast_node& fragmentDefinition)
-{
-	_fragments.emplace(fragmentDefinition.children.front()->string_view(),
-		Fragment(fragmentDefinition, _variables));
-}
-
-} /* namespace graphql::runtime */
+} /* namespace graphql::generator */
